+++ mov/src/org/mov/parser/expression/AbsExpression.java
@@ -34,10 +34,10 @@ public class AbsExpression extends UnaryExpression {
         super(number);
     }
 
-    public float evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) 
+    public double evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) 
 	throws EvaluationException {
 
-        float number = get().evaluate(variables, quoteBundle, symbol, day);
+        double number = get().evaluate(variables, quoteBundle, symbol, day);
 
         return Math.abs(number);
     }
