+++ mov/src/org/mov/parser/expression/DayExpression.java
@@ -33,7 +33,7 @@ public class DayExpression extends TerminalExpression {
         // nothing to do
     }
 
-    public float evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) {
+    public double evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) {
         TradingDate date = quoteBundle.offsetToDate(day);
         return date.getDay();
     }
