+++ mov/src/org/mov/parser/expression/NotExpression.java
@@ -32,7 +32,7 @@ public class NotExpression extends UnaryExpression {
 	super(arg);
     }
 
-    public float evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) 
+    public double evaluate(Variables variables, QuoteBundle quoteBundle, Symbol symbol, int day) 
 	throws EvaluationException {
 
 	if(get().evaluate(variables, quoteBundle, symbol, day) >= Expression.TRUE_LEVEL)
